+++ crates/lunas_generator/src/transformers/js_utils.rs
@@ -233,7 +233,7 @@ pub fn search_json(
                     };
                     let is_target_object = if let Value::Object(object) = object {
                         if let Some(Value::String(object_value)) = object.get("value") {
-                            object_value == "Lunas"
+                            object_value == "Livon"
                         } else {
                             false
                         }